[REF] web_studio: correctly delete a node in report
There was an issue with the 'Delete' button in the report editor sidebar. The
event handler was not correctly bound (it was bound on the whole `card-body`).
Clicking next to it was thus triggering the node deletion.

It now has been refactored as a widget event (the `node` is set in `data`) as it
should have previously been done.