[FIX] web_studio: unset the `widget` on a field
In the report editor, it was not possible to unset the `widget` option
on a field (the `view_change` event was not triggered).

This use case was previously handled by a `console.warn` and a `return`.

The option is now correctly unset.
Note that to be thorough, all other options should be unset when `widget`
is unset as they usually depend on it.