[FIX] web_studio: drop a text in existing col
Before this rev. a traceback occured when dropping a text in an existing
col in a report.

The `_createContent` function was expecting the option `contentInStructure`
to be set on each building block, which is not always the case (if the content
is the same wherever it is dropped).

The mentioned function thus now fallback on `content` in this case.