[FIX] web_studio: correctly show active views
In the Studio Action editor, some views were not correctly shown as active.
It was the case when the view was set in ir.actions.act_window `views` but not
in `view_mode`.

The correct place to find the active view types is in the action `views` key,
set in the ActionManager.

The same fix has been applied for the Studio submenu.

Note that `views` is processed in the ActionManager which transforms it from
something like [['list', 13], ['form', 42]] to an more complex structure
[{type: 'list', ...}] when the action is executed. When modifying the action or
navigating through Studio, the same processing must be done.