[REF] stock_barcode: client action: handle nomenclature
This commit allows users to scan barcode storing product information and
weight thanks to the nomenclature.
Example : ice-cream barcode is 2136478000001. The 5 zeros at the end are
reserved to store the weight of this product. When we scan 2136478124530
in the client action. A new line is added with product = ice-cream and
quantity = 12.453.

task: 1872345