[FIX] web_studio: avoid sidebar flickering
Before this rev. you could observe the sidebar flickering when modifying a node
with a FieldSelector in its component.

There was actually two issues: the report editor was not waiting for the components
deferreds (`concat` does not modify the array in place) and the sidebar was attached
too soon in the DOM (the components `start` functions were called after `appendTo`
of the sidebar in the DOM).

The sidebar is now appended in a fragment before being appended in the DOM.