[FIX] web_enterprise: no flicker when clicking on app name
Since rev. 3d6c7b9, it flickered when the user clicked on the app
name (left part of navbar) when being in an app. The app switcher
briefly appeared while the first action of the app was loading.

This was due to a <div> changed to an <a> with href="#", but the
click event wasn't preventDefaulted and thus redirected to the
home menu.

Task 1878254.