[FIX] web_studio: prevent the user from breaking all reports
With the report editor, it was too easy to break all reports ; remove the `main`
node on the view 'web.report_layout' was quite annoying.

To prevent a user from doing such unwanted mistake, a blacklist of non-editable
template views (the same blacklist used when duplicating a report) is used in
the report edition.

If the user tries to customize on of these views, the operation will be undone
with a warning.