[FIX] stock_barcode: client action: put in pack and multiple page
Instead of applying the pack to the scanned lines, apply it to all the
processed lines as it is done in the onchange of the original barcode
integration. It is needed since the scanned lines are wiped when
changing pages because we consider the flow as ended in this case.