[FIX] web_cohort: appropriate headers of cohort tables when comparing
In comparison mode, headers are supposed to be placed above the tables
containing data and comparison data if at least one table is not empty.
In one configuration (comparison data but no data) a header was missing.

This commit fixes that situation and adds a test of the cohort view
structure.