[FIX] web_grid: keep grid anchor when changing search domain
When navigating with the grid arrow to display previous
weeks, if the user changes the domain in the search view,
he will be bring to the 'today' span of the grid view,
which might be upsetting.
This commit makes the grid view keep the current span,
and provides a test.

Task #1878190