[IMP] web_studio: report-editor: add component building blocks preview
In the previous version of the "add" tab of the report editor, the
building blocks where very generic and their hooks were added while they
were being dragged over the preview zone of the report.

After this commit, the building blocks are more specific and more numerous.
They also behave differently: now when the user hover the building block,
we immediately show on the report preview where they can be dropped,
without modifying the layout of the report. While the building block
is being dragged, we keep on computing he nearest hook to be dropped,
and while the drag is ongoing, it is acceptable to modify the report
preview to reflect the change that will happen.

Also fix multiple creations of elements on drop inside a table (it was
creating as many drop command as there were rows in the table)