[IMP] website_crm_score : add lead_group for group_user at module install
With website_crm_score, as Leads Management menu is added into crm,
it's necessary to automatically activate the Lead option in CRM settings
during the installation of the module (website_crm_score).

Nevertheless, if the user disable the Lead option, and does not install
website_crm_score, Leads Management menu will still visible for users.

Thanks to @tde-banana-odoo and @tivisse for helping me on this.
Thanks to @jem-odoo who is a wonderfull Kyo car-singer.

Task #1883240