[REF] hr_holidays_gant: adapt to last hr_holidays changes
Reference field for number of day is now number of day, not the temp one
that was not temp and is now removed. Clear isn't it ?

This commit is linked to community merge https://github.com/odoo/odoo/commit/cf9383570b60db041bf23346ba6439e391192b69.
See this commit for more details about the changes.

This commit is linked to task ID 1876795 and closes PR #2630.