[FIX] website_sign: improve image alpha edge issue
On some PDF renderer (eg. on evince, gimp, google chrome, google drive
but not on firefox or inkscape) when an image is being rendered there
can be an uninented line on the edge between transparent and not
transparent zone.

The issue is that when calculating the weighted average of color between
neighboring pixels, the color of a totally transparent pixel is taken
into account. By default that color is black and on google chrome it
can't be changed.

So if we for example have a white image of 400x200 inside a signature
area of 600x200, we will have two 100x200 margins (containing
transparent black pixels). On the edge between the white and the
transparent a grey line may be shown on these conditions:

 - faulty PDF renderer
 - the weighted average has enough percent of transparent black
 - changing the DPI/zooming to change this weighted average
 - possibly other heuristics of the PDF renderer

With this commit, we set all black transparent pixels to white
transparent, thus the issue doesn't happen when we have white image on
white PDF document, but inversely will append when having a black
image on black PDF document (which is less likely to happen).

opw-1879688
closes #2620