[FIX] account_reports: fix multicompany ir.rules
The multi-company feature in account reports doesn't work since
odoo/odoo@93cacd5db7ddab

because the applied ir.rule is
['|',('company_id','=',False),('company_id','child_of',[user.company_id.id])]

... that nullify the multi-company domain.

Because the ir.rules are also applied when browsing/searching, the only way to fix it
is to adapt the multi-company widget to this rule.

After this commit, the user will have only access to its current company plus its children in
the accounting reports.