[REF] stock_barcode: client action: matching virtual id <-> id
Use a dummy field and a unique method to get/state the state in order to
know exactly which line is created. This allows to perfectly match the
virtual_id and id and not the first line matching the description and
it's done in order to remove the calls to _endBarcodeFlow (new spec
wants to be able to scan/add product/edit line and scan a destination
location for example).

task 1872345