[IMP] web_grid: check access action for grid view
Currently grid view does not hold any access right check. This means people
not having the right to create records have the 'Add a line' button that
opens a form view.

As the form view is correctly tagged with access rights it is actually
not possible to create a record; however people should not see the 'Add
a line button' at all.

This commit adds the access rights check for grid view in field view get
post processing.