[FIX] stock_barcode_mobile: tap to scan button not clickable
(1) Before this commit, z-index and opacity properties were set on
barcode image. But stock_mobile_barcode button is an absolute
positioned element, so stacking context has been changed because
of the opacity of the image.
For ref: https://css-tricks.com/almanac/properties/o/opacity/#article-header-id-1
Now, we simply removed opacity property on this image.

(2) We also had to transform <a> to <button> because BS4 apply black
background color to non href anchor tag.