[IMP] stock_barcode: package/lot bypass source
Allow the user to directly scan a package or a lot/serial
instead of a source location. In this case the page's default
source location is set in the same way than a manual scan from
the user.

Prevent the user to scan a package that is in a different sublocation
than the current page.

Fix the scan lot if the user firstly scan a lot. In this case
scannedLines was empty and the id was undefined. It then search a
lot for a random product in the picking.

Technically, it does not reject from the source step if the barcode is
not a product. Instead it set the source location and call _step_product
that already allows to bypass the product for a lot or a pack.

task-id: 1872345