[FIX] website_contract: fix singleton error
The code was assuming `_cart_find_product_line()` would return only one record
but it is not (see `_cart_update()` in website_sale's sale_order.py).

This commit fixes the singleton error the same way it is in the original
`_cart_update()` by taking only the [0] element.

Note: This bug should happen only in a rare flow as `_cart_update` is only
      called by the frontend ecommerce and the bug occurs when there is more
      than one time the same product in the SO.
      It will happen if the frontend cart is update in the backend.