[ADD] l10n_co_edi: Colombian electronic invoicing through Carvajal
This module will generate XML files for Colombian invoices upon validation.
These XML files are sent to a 3rd party named Carvajal (company with a role similar to Mexican PAC) and they will process and submit this info to the government.

This module also includes a server action (on invoices) that checks the status of these submitted invoices.
If they are approved by Carvajal it downloads a ZIP containing an official PDF and XML file and posts this in a note on the invoice.

Was PR #2617
Was task #1874136