[FIX] web_enterprise: fold menu on small devices
When there are too much items in the menu (navbar), some of them
are folded into a dropdown (+). Before this rev., this only worked
on medium devices (and higher).

This rev. makes it work for small devices as well.

For extra small devices, the navbar is completely custom so this is
not necessary.

Task 1878254