[FIX] web_studio: do not always load studio bundle
Before this rev. the style used in the report iframe for Studio customizations
was always applied when rendering a report (it was put in web.report_assets_common).

Now the corresponding style has been put in a separate bundle which is only
called when rendering the report using Studio.

Task 1878211