[IMP] stock_barcode: save UI on ViewWidget
When the user click on scan line or edit and get back to the lineWidget,
the highlocation source/dest and the scan message were lost.

This commit save the line widget state before rendering the ViewsWidget
and once the lineWidget is reloaded it will use the saved ui state in
order to render it.