[FIX] web_studio: report_editor: fix drop in table tests
This test was skipped because it was failing on runbot but not in local.
It was failing because of the resolution of chrome headless (1024x764)
was making some absolute position selection to fail because a difference
of a few pixels makes the selection of the drop zones different depending
on the size of the columns.

The fix consisted of fixing the size of the table on which the drop
occurs and set a size small enough to be completely inside the viewport
of the test setup.