[FIX] stock_barcode: scan a locatio without multi locations enabled
To reproduce the issue, don't enable the multi location and scan a
location in a receipt. It results in a stack overflow error.

It is due to the recent changes in the flow (allow to scan lot for
different products, packages etc).

We fix the issue by guarding the bypass of the destination step.