[FIX] account_predictive_bills: fix account prediction using partner_id
Problem: we replaced the search on partner name using the id of the partner prefixed with 'p--' however when we execute to_tsvector('p--123') this will results in two words 'p' and '-123'
select to_tsvector('p--123');
  to_tsvector
----------------
 '-123':2 'p':1

This can result in some wrong prediction as it might match on the single p letter for partners that have yet to create a bill.
To fix the problem, we removed the '-' and as some id might be negative due to migration, we also had to replace the negative sign by the letter x in the query.
example: -32 => x32

Finally instead of using p followed by the partner id, we changed it to 'partnerid' followed by the id. This is done to limit errors for people that might write bills for parking (example: p1 refering for parking slot 1 and might be confused with partner with id 1).