[FIX] mrp_workorder: register lot/sn to component
The quality check type 'register lot/serial' have the purpose to create
a step that record a component of the BoM. However it could be use in
order to register a not tracked component. So the step name does not
reflect the reality and could hide the functionality to the end user.

The previous name 'register component' seems more appropriate.