[FIX] web_studio: correctly create blank report layout
Before this fix, when creating a report with the blank layout the html
and assets were duplicted inside the main element of the report itself
because of the differences between the template basic_layout (that
already called the html_container) and the other templates (that do not
call themselves the html_container).

In thix fix, when creating a new report with the blank layout, we do not
call html_container.