[FIX] web_studio: fast drag on large reports always creates hooks
Before this fix, the hooks were added only when the mouse was inside a
building block, and before the drag starts.
It works well when the report is of a reasonable size and the user is
not too fast with the mouse.
In the normal case, the events happen in this order:
1. mouse enter in building block --> start adding the hooks (might take long)
2.                                   <----------- clean hooks
3.                                   I am done computing hooks
4. begin drag --> set isDragging to true, when it is true, no cleanup of hooks is allowed
5. dragging --> computing nearest hook
6. mouse leaves the building block to go to the preview report
                                    <----------- clean hooks do nothing
7. dragging...
8. dropping --> add the component set isDragging to false
9.                                  <----------- clean hooks

In the case where computing the hooks takes a long time, and the user is
fast with the mouse, the order of the events were broken and the hooks
were removed because the mouse was leaving the building block before the
drag event was exectued. In this case the sequence of events looks like
this:
1. mouse enter in building block --> start adding the hooks (might take long)
2.                                   <----------- clean hooks
3.                                   I am done computing hooks
4. mouse leaves the building block to go to the preview report
                                    <----------- clean hooks
5. begin drag --> set isDragging to true, when it is true, no cleanup of hooks is allowed
6. dragging --> computing nearest hook
...

After this fix, if no hooks exist on the report when the begin drag event
occurs, we will add them a second time.