[IMP] quality_iot, quality_mrp_iot: pedals
Pedals, more specifically foot switches work like
keyboards.  We listen to the IoT Box to see which keys
are tapped and based on the information on the workcenter
(in the boxes field) we trigger the action that needs to be done.
The actions defined are the same as in the barcode_trigger defined
by the barcode modules: so it will click on the visible button with
the action as barcode_trigger.

Only one tab can own a pedal and if a tab is blocked,
the user can still take ownership by pushing the button.