[ADD] iot_drivers : IOT Box drivers
This is the module we use to put the IoT drivers.  You can however
put them in any module.  They will be downloaded to the IoT Box if they
are in the drivers subdirectory.

We provided 3 drivers to start with: the Sylvac USB and Bluetooth calipers
and the keyboard driver.  They import the usb and bluetooth managers
provided in the hw_drivers module where they will reside on the IoT Box.
It is important that this import starts with odoo.addons and not just addons,
otherwise the module will be loaded twice