[FIX] stock_barcode: split common packages lines
Issue 1:
- Scan a product
- Scan a pack that contains the previous product.

-> traceback because package_already_scanned compare
barcode line packages with current package. However barcode lines
do not have always a package. This commit

Issue 2 (without 'move entire pack'):
- Create 2 packs that contains the same product
- Scan pack 1
- Scan pack 2

-> barcode line with the same product but with a different package are
merged together. It should not be the case since we want to see the
source and the destination package for each line. It happens because
_findCandidateLineToIncrement do not care about package.

This commit add a package check when the system check if a move line
could match the barcode action. It also ensure that a barcode line have
a package before check for it.