[FIX] web_studio: column hooks in table-bordered
Before this fix, the displayed hooks for 'Field Column' block were breaking
the table layout.

This is due to rules applied on .table-bordered tables in web_editor.common.scss
which basically apply `table-layout: fixed` (among others) ; the additionnal
columns created for hooks thus take extra space.

This commit thus reverses these rules, until they are removed (which cannot be
done at the moment because the display of tables on the website depends on them
so another solution needs to be found).

Task 1878211