[ADD] iot: basic models + controllers server-side, print reports, measure widget
When an IoT box is configured as IoT box, it will send all the information
about itself and its devices to its server. This way the server knows which ip
and which devices are on the iotbox. Reports and quality points will be linked
to devices. In the module iot, the models and the controller are defined for this.
There is also a javascript part with a widget for taking a measure and the possibility
to link any report with a printer device.