[IMP] iot: connection wizard and more
We wanted to ease the registration and authentication of the different
IoT boxes on the Odoo server by implementing a connect
wizard.  (by a connect button replacing the create button)

A token is created when launching the registration wizard if none exists or if
it is old.  Either the token is passed through the network or copy pasted
to the access point web interface in order for the IoT Box to register
itself with the server.

In the wizard, either you copy paste the token or you press the scan button to
check in the network.  If it can use webrtc to find it own ip address, it will
scan the 256 addresses in that range, otherwise it will fall back on the domains
the point_of_sale uses.

- A route was added to download the drivers found on the server to the IoT Box.
- iot.trigger model is added to support pedals (in the work orders for now)
- A computed url to go immediately to the config page of the box
- 'Take Measure' separate button widget instead of FieldFloat (thanks to Pierre (pha))
so the button is immediately next to the validate button and is not placed relative
to the fieldfloat resulting in a cleaner interface

Thanks to qle and twa