[FIX] delivery_dhl: rename app to DHL EXPRESS
DHL is a big services and offer different services aroud the world,
for example:
- DHL Express: the generic worldwide courier service
- DHL Parcel: a specific parcel service
- DHL Paket: a German-centric service from the German post office

This module only works for DHL Express other services have their own
specific API and this module do not implemented them.

This commit rename app name and description in order to specify that
this app can only be use for DHL Express services.

task-1888731