[IMP] sign: do not pre-fill signature box with user signature
Currently when a customer linked to an user fills its signature / initials
box all next documents are already filled with its signature / initials.
This behavior can be confusing to users because they could think the
document is already signed which is not the case.

In this commit we update that behavior by uploading the signature / initials
when loading the signature box instead of pre-filling it in the document
itself. That way users can quickly sign using their signature but still see
an empty box while it is not done.

Co-Authored-By: Yannick Tivisse <yti@odoo.com>
Co-Authored-By: Romain Libert <rli@odoo.com>