[IMP] l10n_mx_edi: Hide payment_method_id in Mexican payments
In the payments without MX localization, the payment method could be
Manual, Check, Electronic, etc. and are not related with the Mexican
payments method (PUE, PPD).

Then if We print in the PDF example:

Payment Method: Manual

The Mexican partners have a confusion by this value.

Note: Was added double if to keep the same div that add Odoo to avoid
errors if the xpath is inherit from other module.

closes odoo/enterprise#2742