[FIX] web_studio: correctly size the footer in editor
Before this rev. the footer in the report editor didn't have a correct width.

This is due to the fact that it's put in position fixed to push it in the bottom
of the page.

The width is now dynamically set according to the .page width.

Achieving this wasn't possible with static css rules (without changing the whole
report style) because the fixed element width is not linked to its relative
parent anymore but to the document.