[FIX] quality_mrp_iot: make sure measure is saved in work order view
We wanted to make the measure field readonly on the
work order view when doing a quality step of type
measure and there is an IoT device associated.

This gives problems as the value won't be saved
and it is better to make it writable as there
can always be a connection problem or it might be
easier to type it manually.