[IMP] iot: flexible urls for http / https
In order for the IoT Box to work with the SaaS
that is in https, it is best for the browser
to use the same protocol to the IoT Box as to
the server. Otherwise in https, it will give
an error trying to connect in http to the IoT Box.

For this we check whether the base url config parameter
is in http/https and use the same protocol for the IoT Box
in the js.