[FIX] stock_barcode: correct highlightning process
The current scanned line is highlighted in green (the reserved quantity in not exceeded) or red (the quantity done is higher that the reserved one). The color choice is made before incrementing teh quantity. Is was working well before as we always scanned quantity of 1. In case of quantity > 1, the line could be green while it should be red.