[FIX] web_studio: allow to search contact widget fields
Before this rev. a traceback occured when clicking on 'Search More' on the
special `many2many_select` widget used to set the 'Displayed field' option
on the contact widget in a report.

The widget hadn't correctly handled this use case in the internal m2o used
in the many2many tag widget.

Two things have been done to fix this issue:
 * the internal m2o limit has been set to the selection limit (which is quite small)
 * the m2o `autoCompleteSources` has been correctly implemented, to avoid overriding
   the call_service

closes odoo/enterprise#2815