Revert "[FIX] sign: Remove useless lib imports"
This reverts commit bd629e8778bc2cc7ce7befa09af868f772d16e3e.

The goal was to get rid of the "Uncaught ReferenceError: jQuery is not defined"
when editing a pdf. The side effect of this commit was also to remove the
resizable box feature.

This revert has for goal to bring back the feature while waiting to
find a proper solution about the error in the JS console.