[FIX] mrp_mps: wrong qty available in test
It happens due to commit 960360afe478a8f7b9c456721b5591154952a37d that
slightly modify usage of qty_available with a to_date in context. Before
the commit it was only suppose to work with a Date and after it works
with a Datetime.

Since test pass a date with only years, months and days it will use 0 as
default for the hours and minutes. So it will not find today's moves.
(MPS will skip context if to_date < today)

This commit add hours and minutes.