[IMP] l10n_mx_edi_external_trade: Use the Odoo field to incoterm
Now Odoo add the incoterm field in account module, We do not need add
this field again.

Removed field in the view and change references to use the Odoo field.

Was added a TODO to remove the field on master and was changed the label
to avoid the warning:
```Two fields (l10n_mx_edi_incoterm, incoterm_id) of account.invoice() have
the same label: Incoterm.```

**Extra** fixed the test to use the correct model

closes odoo/enterprise#2768