[FIX] web_studio: display Studio menu after menu creation
When creating a new menu through the 'Edit menu' link in Studio, the Studio menu
(the second navbar beneath the main menu) wasn't displayed anymore.

In the normal flow, this sidebar is rendered (via a bus event `studio_main`) in
the view editor action (see main.js). When creating a new menu, this action is
not reloaded, but the menu is still destroyed and reinstantiated. The Studio menu
thus needs to also be reinstantied when reinstantiating the whole menu when
main is the current mode.

Note that while adding a step in the test tour, a wild bug suddendly popped up.
When clicking twice on the button to add a new menu, two menus were actually
added. The button is now disabled (so we can click on the first modal 'Confirm'
button and on the second modal same button in two consecutive steps).

Closes #2822