[FIX] l10n_mx_edi: Refactored the payment details by invoice.
Now the SAT was improved the validations on the amounts in the invoice
details and the amount in the total.

Fixed cases with multicurrency. Now the rate is calculated in base to
the amount paid in the payment / sum(amount paid by each invoice), now
is not considered the rate in Odoo, because the amounts must make match
in the CFDI.

Also was improved the 'Serie & Folio' attributes to express the same
value that is found in the invoice.

closes odoo/enterprise#2848