[FIX] account_reports: apply currency conversion on debit and credit amounts in trial balance
Before that, amounts reported in 'debit' and 'credit' columns of the trial balance were treated as if they had been expressed in current company's currency from the start. This was wrong in multi company environments with companies using distinct currencies, when displaying the report for multiple companies at once, as amounts that should have been converted were kept as-is, simply changing the symbol of the currency. With this fix, we now convert everything to the user's current company's main currency.

This bug was spotted thanks to ticket 1888360.