[FIX] mrp_mps: to_supply quantity double
It happens in v12 due to the refactoring on datetime fields.
In v11 when accessing a datefield from a record it returns a string,
v12 it returns a datetime.date object. However in both version the js
send a string to the server. Thus in v12 it compares a string with a
date object and it goes in the 'else' that will create a new forecast
object. However the code that create a new forecast object never works
since its creation because it creates the new forecast object with the
quantity exisiting and not the needed quantity in order to obtain the
quantity introduced by the user (that's why it double it).

This commit transfor the string in date object before comparaison and
correct the else in order to introduce correct quantity.

opw-1891909

closes odoo/enterprise#2835