[FIX] documents: add attachment type on write
Changed the write of attachments so the type is set based
on what's written on the attachment.

Reason: there are multiple places where the data of a file can(/could)
be changed (eg: the form view) and the current behaviour kept
the old type of attachments if the data wasn't changed through the
Documents dedicated button.

This change makes so the type is set by the backend in a generic way
instead of in a specific place of the front-end.