[IMP] l10n_mx_edi_payment: Return 2 False always that is to MX
When is returned to draft a payment must be returned to False, this
to allow recalculate this values when the payment is validated again.

But the filtered was incomplete, because when the payment is nos stamped
and is returned to draft also must be reset that values.

Also was improved that use `write` in all records and not update by each
record.