[IMP] l10n_mx_edi_external_trade: Support countries without state
Some countries do not have states (Taiwan for example), and in this case
the attribute "Estado" to the receiver do not have a value to set.

In this case, the attribute "Estado" cannot be empty because that is
required, in that case could be used other string to avoid the empty
value. (NA) in this case.

The SAT only validate the value in the attribute when the country code
is MXN, USA or CAN.

Here the SAT documentation:
 - Si la clave de país es {ZZZ} o la clave del país no existe en la
 columna c_Pais del catálogo c_Estado, se podrá registrar texto
 libremente.

closes odoo/enterprise#2860