[FIX] account_reports: fix financial reports in case of XLSX export
When exporting in XLSX, there's no formatting and the raw value is writen in the file. So in Excel, this resulted with some cells showing 'almost 0' amounts instead of 0 because of the python float representation.
Also, returning a False when expecting a 0 was transformed in the XLSX as the formula '=False()', which wasn't very clean.