[FIX] account_reports: dates support in XLSX export
Dates and times in Excel are represented by real numbers, for example “Jan 1 2013 12:00 PM” is represented by the number 41275.5.
The integer part of the number stores the number of days since the epoch and the fractional part stores the percentage of the day.

So when writing a date in the XLSX, we need to give proper format or it will be shown as a float in the Excel file.