[FIX] account_reports: fix cash flow statement
Latest refactoring made in https://github.com/odoo/enterprise/commit/ddfa846e changed the behavior of the cash basis option on reports making it only an option to toggle the visibility of the cash basdis filter) and, now the cash flow statement don't have to have this option set to True anymore.

So the query's with_statement have to be adapted to use the debit/credit/balance fields as their cash_basis variants won't be displayed at all by the report.

closes odoo/enterprise#2871