[FIX] account_yodlee: fix some errors with new synchronization system
For some reason, yodlee API is not entirely the same between sandbox and production, resulting in the status being an integer (1-2-3) in production as opposed to a string (SUCCESS-FAILED-ACTION_ABANDONNED) in sandbox. This result in a success message being always displayed when trying to synchronize even if an error has occured.

Also fix problem where when an error occured no message was sent to the chatter of the online synchronization record.

Better display of error message in case the server is down (or no internet connection)