[FIX] documents: display rules of reloaded record
Before this rev., in the DocumentKanbanView, the rules weren't
correctly displayed (in the Inspector), when the corresponding
record had previously been reloaded individually (for instance,
after a button on its card, like the priority star, had been
clicked).

This was due to a wrong configuration of the m2m field given to
the BasicModel, leading to the related fields no being fetched
when the record was reloaded. This whole fieldsInfo logic is too
complicated, and we should try to put some effort to clean it in
the future.

Task 1889245

closes odoo/enterprise#2880