[FIX] documents: tour display tooltip ball
The tooltip balls are inserted at next to the DOM elements it points and
not just absolutely so it follows the scrolling.

In the documents tour, one ball was hidden it was getting outside its
container, so this commit just put it below.

Do note that the tooltip shown on hover does not have this issue since
in this instance we position it absolutely in the body (so this is
always shown but doesn't scroll with the element).

opw-1893035
closes #2888