[FIX] project_forecast: Empty "Assigned to" field raises an error
For each change of the "assigned to" field, an on-change modifies the
number of hours an employee work according to his time zone. The problem
is, when the field is empty, it fails to compute the timezone of nobody.

This PR corrects that behavior by checking the field is not empty.

opw-1895271

closes odoo/enterprise#2889