[ADD] l10n_mx_edi_tax_cash_basis: amount computation
Hooking method to allow l10n_mx to recompute `rounded_amt` var used to
fetch debit & credit recorded in Tax Cash Basis Journal Entries.

Main
-
By overriding `_get_amount_tax_cash_basis` method in
`account.partial.reconcile` We are allowed to passed down to
`create_tax_cash_basis_entry` method the values - that according to
l10n_mx - fit with the proper ones for debit and credit in Tax Cash
Basis Journal Entry.

Explanation
-
In l10n_mx the Tax Cash Basis Journal Entries - when dealing with
multi-currency - the debits and credits must be computed at Payment Date
rate instead of at Invoice Date rate as done in core module.

Be aware
-
Do not be mislead by the fact that`rounded_amt` var is the one taken to
do those computations. It is just a mean to the final goal. Rounding is
not an issue and has never been.

Dependencies
-
This PR leverages on commit
https://github.com/odoo/odoo/commit/9d2d539a1631bc15516c75a100d8e0f305df2cd6

Background
-

By allowing this merge to flow we will be able to address the following
issues:
- https://github.com/odoo/odoo/issues/25057
- https://github.com/odoo/odoo/pull/17196
- https://github.com/odoo/odoo/issues/9972

Final Goal
-
Being able to perform the computation of the debits and credits
according to l10n_mx

Closes #3137

closes odoo/enterprise#3256