[FIX] delivery_ups: change carrier on picking
- Create 2 UPS delivery methods
  UPS 1: UPS Service Type = UPS Standard
  UPS 2: UPS Service Type = UPS Ground
- Create a SO and choose delivery method UPS 1, validate
- On the picking, change to delivery method UPS 2, validate

The UPS Service Type which is sent to the UPS web service is the one of
UPS 1, not UPS 2.

This is because `picking.ups_service_type` is a realted to
`sale_id.ups_service_type`. Therefore, we always select the UPS Service
Type of the SO, whatever is chosen on the picking.

To fix this, we give the priority to
`picking.carrier_id.ups_default_service_type`.

Closes #2825

opw-1871563

closes odoo/enterprise#2896