[FIX] web_studio: use `address` class in corresponding block
When dropping an 'Address' block in the Report designer, the corresponding HTML
was basically a field in a row/col.

However, when using snailmail, an element with the `address` class is required in
the report to determine the address position.

This class is now correctly added when dropping the block.

closes odoo/enterprise#2899