[FIX] account_reports: don't indent numbers when grouping is enabled
When grouping with a filter the indentation is confusing. You end up
with e.g.:

                       COMPANY A   COMPANY B

ITEM A                   $ 314       $ 159
ITEM B                   $ 265       $ 358
  ITEM C                   $ 979       $ 323
  ...
    ITEM D                   $ 846       $ 264
    ...

The numbers are indented just like the row names. This makes it hard
to read. This commit aligns numbers in a column by disabling the
margin:

                       COMPANY A   COMPANY B

ITEM A                     $ 314       $ 159
ITEM B                     $ 265       $ 358
  ITEM C                   $ 979       $ 323
  ...
    ITEM D                 $ 846       $ 264
    ...

opw-1896519