[FIX] helpdesk: original message in portal
- Set-up an alias to create a ticket from an email
- Send an email with a subject and a body
- Access the portal view of the ticket /helpdesk/ticket/<id>

The description of the problem which was included in the body is not
displayed anywhere. This is confusing for the end user who cannot find
back the description of his problem.

By default, `website_message_ids` only displays `comment` messages, not
`email`.

Similar to https://github.com/odoo/odoo/commit/c574490a66038618214b3bf6

opw-1892151

closes odoo/enterprise#2914