[FIX] account_online_sync: UI components not shown on sync failure
When a bank synchronization fails, an error message should appear
along with a link to a detailed error report. Those components
weren't shown because the xpath id has been ghosted by another div.

This PR moves the two components defined in the xpath in another
one that is also related to bank operations and that is correcly
shown on the UI.

opw-1892525