[FIX] web_enterprise: always wrap o_search_options
div.o_search_options contains the search filters and would overflow in
some cases:

1. make your browser window narrow (<= 500px)
2. navigate to Account > Reporting > Profit and Loss
3. refresh the page (F5)

The search filters would nicely wrap before the refresh, but after
refreshing a vertical scroll bar appears. The filters no longer wrap
and part of them appear outside the Odoo layout.

It's because the filters sit inside a .btn-group which is a flex
container (display: inline-flex). Flex containers don't wrap by
default.

It happens to work before the refresh because the ControlPanel's
update function is called with options.clear = true, which results in:

this.nodes.$searchview_buttons.show();

which adds a style="display: block;" attribute to the
.o_search_options div. This stops the div from acting as a flex
container.

This fixes the issue by always considering this div as display: block.

opw-1896519